Add a lower bound to --storage.tsdb.max-block-chunk-segment-size

+++ cmd/prometheus/main_test.go
@@ -238,12 +238,12 @@ func TestWALSegmentSizeBounds(t *testing.T) {
 	}
 }
 
-func TestMaxChunkSize(t *testing.T) {
+func TestMaxBlockChunkSegmentSizeBounds(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping test in short mode.")
 	}
 
-	for _, size := range []string{"1MB", "512MB", "1GB"} {
+	for size, expectedExitStatus := range map[string]int{"512KB": 1, "1MB": 0} {
 		prom := exec.Command(promPath, "-test.main", "--storage.tsdb.max-block-chunk-segment-size="+size, "--config.file="+promConfig)
 
 		// Log stderr in case of failure.
@@ -257,13 +257,25 @@ func TestMaxChunkSize(t *testing.T) {
 		err = prom.Start()
 		require.NoError(t, err)
 
-		done := make(chan error, 1)
-		go func() { done <- prom.Wait() }()
-		select {
-		case err := <-done:
-			t.Errorf("prometheus should be still running: %v", err)
-		case <-time.After(5 * time.Second):
-			prom.Process.Kill()
+		if expectedExitStatus == 0 {
+			done := make(chan error, 1)
+			go func() { done <- prom.Wait() }()
+			select {
+			case err := <-done:
+				t.Errorf("prometheus should be still running: %v", err)
+			case <-time.After(5 * time.Second):
+				prom.Process.Kill()
+			}
+			continue
+		}
+
+		err = prom.Wait()
+		require.Error(t, err)
+		if exitError, ok := err.(*exec.ExitError); ok {
+			status := exitError.Sys().(syscall.WaitStatus)
+			require.Equal(t, expectedExitStatus, status.ExitStatus())
+		} else {
+			t.Errorf("unable to retrieve the exit status for prometheus: %v", err)
 		}
 	}
 }
